% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar.R
\name{c5_import}
\alias{c5_import}
\title{Import data from export CANREG5 rename columns and check for presence of esential variables.}
\usage{
c5_import(dir = file.choose(), sep = ",", encoding = "cp1252",
  delim = ",")
}
\arguments{
\item{dir}{chr: Directory path. Default file.choose.}
}
\description{
Import data from export CANREG5 rename columns and check for presence of esential variables.
}
