% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportaciones.R
\name{its_export}
\alias{its_export}
\title{Export data for check with IarcTools.}
\usage{
its_export(df, dir = file.choose(), type = "pm")
}
\arguments{
\item{df}{data.frame: data imported and transformed.}

\item{dir}{chr: Directory path. Default file.choose().}
}
\description{
Export data for check with IarcTools.
}
