% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar.R
\name{read_mul}
\alias{read_mul}
\title{Read data from .mul file from IarcTools, rename variables and clean data.}
\usage{
read_mul(file = file.choose())
}
\arguments{
\item{file}{chr: Directory file.}
}
\description{
Read data from .mul file from IarcTools, rename variables and clean data.
}
